refactored the code to add separate folder for each product/app
